requirements.txt
streamlit
python-dotenv==1.0.1
pydantic
pydantic-ai
groq
langchain
langchain-community
sentence-transformers
torch
faiss-cpu
langchain_huggingface 
rich==13.4.0
fastapi